fix(aero-storage): avoid debug panics + 32-bit truncation in VHD/QCOW2
- Remove the `debug_assert` in VHD read path and make the per-block read loop use
  checked subtraction to avoid underflow/wrap hangs.
- Use fallible u64->usize conversions for VHD block index calculations so 32-bit
  builds don’t silently truncate large LBAs.
- Clamp/convert `write_zeroes` chunk sizes safely (VHD + QCOW2) to avoid any
  potential 32-bit `as usize` truncation edge cases.

+++ crates/aero-storage/src/qcow2.rs
@@ -1099,7 +1099,8 @@ fn write_zeroes<B: StorageBackend>(backend: &mut B, mut offset: u64, mut len: u6
     const CHUNK: usize = 64 * 1024;
     let buf = [0u8; CHUNK];
     while len > 0 {
-        let to_write = (len as usize).min(CHUNK);
+        // Convert to usize *after* clamping so we never truncate `len` on 32-bit builds.
+        let to_write = len.min(CHUNK as u64) as usize;
         backend.write_at(offset, &buf[..to_write])?;
         offset = offset
             .checked_add(to_write as u64)
